+++ pyramid_google_login/exceptions.py
@@ -1,0 +1,11 @@
+
+class Base(Exception):
+    pass
+
+
+class AuthFailed(Base):
+    pass
+
+
+class ApiError(Base):
+    pass